Move showStore method
