+++ kDrive/UI/Controller/NewFolder/NewFolderTypeTableViewController.swift
@@ -90,9 +90,9 @@ class NewFolderTypeTableViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         if content[indexPath.row] == .dropbox && (driveFileManager.drive.pack == .free || driveFileManager.drive.pack == .solo) {
             let floatingPanelViewController = DropBoxFloatingPanelViewController.instantiatePanel()
-            (floatingPanelViewController.contentViewController as? DropBoxFloatingPanelViewController)?.actionHandler = { [weak self] _ in
+            (floatingPanelViewController.contentViewController as? DropBoxFloatingPanelViewController)?.actionHandler = { [weak self] _ in
                 guard let self = self else { return }
-                UIConstants.showStore(from: self, driveFileManager: self.driveFileManager)
+                StorePresenter.showStore(from: self, driveFileManager: self.driveFileManager)
             }
             present(floatingPanelViewController, animated: true)
             return
